+++ src/CodeGen/compilation/prolog.rs
@@ -7,7 +7,7 @@ impl CompilationHelper {
     pub fn compile_prolog(&mut self, sink: &mut Vec<MachineInstr>) {
         if self.alloc.stack_off > self.call.shadow(self.arch) {
             let mut instr = MachineInstr::new( MachineMnemonic::Prolog );
-            instr.add_operand( MachineOperand::Imm(self.alloc.stack_off) );
+            instr.add_operand( MachineOperand::Imm(self.alloc.stack_off as f64) );
 
             sink.push( instr );
         }
@@ -17,7 +17,7 @@ impl CompilationHelper {
     pub fn compile_epilog(&mut self, sink: &mut Vec<MachineInstr>) {
         if self.alloc.stack_off > self.call.shadow(self.arch) {
             let mut instr = MachineInstr::new( MachineMnemonic::Epilog );
-            instr.add_operand( MachineOperand::Imm(self.alloc.stack_off) );
+            instr.add_operand( MachineOperand::Imm(self.alloc.stack_off as f64) );
 
             sink.push( instr );
         }
